phase4/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.1",
    "mcp>=1.26.0",
    "openai-agents>=0.12.5",
    "psycopg-binary>=3.3.3",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.37",
]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
]
